константы в модули
