+++ js/check_ins_outs.js
@@ -1,0 +1,7 @@
+const CHECK_INS_OUTS = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+export {CHECK_INS_OUTS};
