+++ js/descriptions.js
@@ -1,0 +1,14 @@
+const DESCRIPTIONS = [
+  'большая',
+  'чистая',
+  'светлая',
+  'просторная',
+  'пентхауз',
+  'ремонт',
+  'прекрасный вид',
+  'нет соседей',
+  'возле метро',
+  'возле парка',
+];
+
+export {DESCRIPTIONS};
