+++ js/features.js
@@ -1,0 +1,10 @@
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+
+export {FEATURES};
