+++ js/titles.js
@@ -1,0 +1,14 @@
+const TITLES = [
+  '1-комнатная',
+  '2-комнатная',
+  '3-комнатная',
+  '4-комнатная',
+  'студия',
+  'кавалерка',
+  'хостел',
+  'VIP',
+  'люкс',
+  'президентский номер',
+];
+
+export {TITLES};
