+++ js/types.js
@@ -1,0 +1,9 @@
+const TYPES = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel',
+];
+
+export {TYPES};
